mcp-gateway/pyproject.toml
[project]
name = "yt-factory-mcp-gateway"
version = "1.0.0"
description = "MCP Gateway for YT-Factory - Google Trends, YouTube Publishing, Analytics"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.0",
    "pydantic>=2.0",
    "httpx>=0.25",
    "google-api-python-client>=2.100",
    "google-auth-oauthlib>=1.1",
    "google-auth-httplib2>=0.1",
    "pytrends>=4.9",
    "python-dotenv>=1.0",
    "structlog>=23.2",
    "tenacity>=8.2",
    "aiofiles>=23.2",
]

[dependency-groups]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.21",
    "ruff>=0.9",
    "mypy>=1.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.ruff.format]
quote-style = "double"
